+++ build.sh
@@ -15,7 +15,8 @@
 provider_directory="$(pwd)"
 GO111MODULE=on GOSUMDB=off go get -d github.com/megaport/megaportgo
 rm -f bin/*
-provider_filename="$(pwd)/bin/terraform-provider-megaport_$(git describe --tags)"
+version="$(git describe --tags)"
+provider_filename="$(pwd)/bin/terraform-provider-megaport_$version"
 provider_filename_no_version="$(pwd)/bin/terraform-provider-megaport"
 go build -o $provider_filename
 cp $provider_filename "bin/terraform-provider-megaport"
@@ -27,7 +28,7 @@ plugin_directory="$(pwd)/.terraform.d/plugins/${arch}/"
 mkdir -p $plugin_directory
 ln -s $provider_filename_no_version $plugin_directory
 echo "Symbolic link created from build directory to terraform.d. < 0.13"
-plugin_directory="$(pwd)/.terraform.d/plugins/megaport.com/megaport/megaport/0.1.0/${arch}/"
+plugin_directory="$(pwd)/.terraform.d/plugins/megaport.com/megaport/megaport/${version:1}/${arch}/"
 mkdir -p $plugin_directory
 ln -s $provider_filename_no_version $plugin_directory
 echo "Symbolic link created from build directory to terraform.d. >= 0.13"