+++ examples/cloud_to_cloud_aws_azure/aws.tf
@@ -33,7 +33,7 @@ data "aws_ami" "amzn2linux" {
 }
 
 // networking
-resource "aws_vpc" "tf_test" {
+resource "aws_vpc" "vpc" {
   cidr_block = var.aws_vpc_cidr
 
   tags = {
@@ -41,9 +41,9 @@ resource "aws_vpc" "tf_test" {
   }
 }
 
-resource "aws_subnet" "tf_test" {
-  vpc_id                  = aws_vpc.tf_test.id
-  cidr_block              = aws_vpc.tf_test.cidr_block
+resource "aws_subnet" "subnet" {
+  vpc_id                  = aws_vpc.vpc.id
+  cidr_block              = aws_vpc.vpc.cidr_block
   map_public_ip_on_launch = false
 
   tags = {
@@ -51,21 +51,21 @@ resource "aws_subnet" "tf_test" {
   }
 }
 
-resource "aws_route_table" "tf_test" {
-  vpc_id = aws_vpc.tf_test.id
+resource "aws_route_table" "route_table" {
+  vpc_id = aws_vpc.vpc.id
 
   tags = {
     Name = "${var.prefix} Terraform Route Table"
   }
 }
 
-resource "aws_route_table_association" "tf_test" {
-  subnet_id      = aws_subnet.tf_test.id
-  route_table_id = aws_route_table.tf_test.id
+resource "aws_route_table_association" "route_table_association" {
+  subnet_id      = aws_subnet.subnet.id
+  route_table_id = aws_route_table.route_table.id
 }
 
-resource "aws_security_group" "tf_test" {
-  vpc_id      = aws_vpc.tf_test.id
+resource "aws_security_group" "security_group" {
+  vpc_id      = aws_vpc.vpc.id
   name        = "${var.prefix} Terraform Security Group"
   description = "${var.prefix} Terraform Security Group"
 
@@ -96,33 +96,33 @@ resource "aws_security_group" "tf_test" {
 }
 
 // vpn gateway
-resource "aws_vpn_gateway" "tf_test" {
-  vpc_id = aws_vpc.tf_test.id
+resource "aws_vpn_gateway" "vpn_gateway" {
+  vpc_id = aws_vpc.vpc.id
 
   tags = {
     Name = "${var.prefix} Terraform Virtual Gateway"
   }
 }
 
-resource "aws_vpn_gateway_route_propagation" "tf_test" {
-  vpn_gateway_id = aws_vpn_gateway.tf_test.id
-  route_table_id = aws_route_table.tf_test.id
+resource "aws_vpn_gateway_route_propagation" "vpn_gateway_route_propagation" {
+  vpn_gateway_id = aws_vpn_gateway.vpn_gateway.id
+  route_table_id = aws_route_table.route_table.id
 }
 
 // direct connect
-resource "aws_dx_gateway" "tf_test" {
+resource "aws_dx_gateway" "dx_gateway" {
   name            = "${var.prefix} Terraform DX Gateway"
   amazon_side_asn = var.aws_dx_gateway_asn
 }
 
-resource "aws_dx_gateway_association" "tf_test" {
-  dx_gateway_id         = aws_dx_gateway.tf_test.id
-  associated_gateway_id = aws_vpn_gateway.tf_test.id
+resource "aws_dx_gateway_association" "dx_gateway_association" {
+  dx_gateway_id         = aws_dx_gateway.dx_gateway.id
+  associated_gateway_id = aws_vpn_gateway.vpn_gateway.id
 }
 
-resource "aws_dx_hosted_private_virtual_interface_accepter" "tf_test" {
-  virtual_interface_id = megaport_aws_connection.example.aws_id
-  vpn_gateway_id       = aws_vpn_gateway.tf_test.id
+resource "aws_dx_hosted_private_virtual_interface_accepter" "dx_hosted_private_virtual_interface_accepter" {
+  virtual_interface_id = megaport_aws_connection.aws_vxc.aws_id
+  vpn_gateway_id       = aws_vpn_gateway.vpn_gateway.id
 
   tags = {
     Side = "Accepter"
@@ -131,11 +131,11 @@ resource "aws_dx_hosted_private_virtual_interface_accepter" "tf_test" {
 }
 
 // instance
-resource "aws_instance" "tf_test" {
+resource "aws_instance" "instance" {
   ami                    = data.aws_ami.amzn2linux.id
   instance_type          = var.aws_ec2_instance_type
-  subnet_id              = aws_subnet.tf_test.id
-  vpc_security_group_ids = [aws_security_group.tf_test.id]
+  subnet_id              = aws_subnet.subnet.id
+  vpc_security_group_ids = [aws_security_group.security_group.id]
   key_name               = var.aws_ec2_key_pair_name
 
   tags = {