+++ examples/lag_port/megaport.tf
@@ -17,29 +17,29 @@
 terraform {
   required_providers {
     megaport = {
-      source = "megaport/megaport"
+      source  = "megaport/megaport"
       version = ">=0.1.4"
     }
   }
 }
 
 provider "megaport" {
-    username                = "my.test.user@example.org"
-    password                = "n0t@re4lPassw0rd"
-    mfa_otp_key             = "ABCDEFGHIJK01234"
-    accept_purchase_terms   = true
-    delete_ports            = true
-    environment             = "staging"
+  username              = "my.test.user@example.org"
+  password              = "n0t@re4lPassw0rd"
+  mfa_otp_key           = "ABCDEFGHIJK01234"
+  accept_purchase_terms = true
+  delete_ports          = true
+  environment           = "staging"
 }
 
-data megaport_location nextdc_brisbane_2 {
+data "megaport_location" "bne_nxt2" {
   name = "NextDC B2"
 }
 
-resource megaport_port lag_port {
-  port_name      = "Test Lag Port"
+resource "megaport_port" "lag_port" {
+  port_name      = "Terraform Example - LAG Port"
   port_speed     = 10000
-  location_id    = data.megaport_location.nextdc_brisbane_2.id
+  location_id    = data.megaport_location.bne_nxt2.id
   lag            = true
-  lag_port_count = 5
+  lag_port_count = 3
 }