+++ examples/single_port/megaport.tf
@@ -17,30 +17,29 @@
 terraform {
   required_providers {
     megaport = {
-      source = "megaport/megaport"
+      source  = "megaport/megaport"
       version = ">=0.1.4"
     }
   }
 }
 
 provider "megaport" {
-    username                = "my.test.user@example.org"
-    password                = "n0t@re4lPassw0rd"
-    mfa_otp_key             = "ABCDEFGHIJK01234"
-    accept_purchase_terms   = true
-    delete_ports            = true
-    environment             = "staging"
+  username              = "my.test.user@example.org"
+  password              = "n0t@re4lPassw0rd"
+  mfa_otp_key           = "ABCDEFGHIJK01234"
+  accept_purchase_terms = true
+  delete_ports          = true
+  environment           = "staging"
 }
 
-data megaport_location ndc_b1 {
+data "megaport_location" "bne_nxt1" {
   name    = "NextDC B1"
   has_mcr = false
 }
 
-resource megaport_port tf_test {
-  port_name   = "Test Port"
+resource "megaport_port" "port" {
+  port_name   = "Terraform Example - Port"
   port_speed  = 1000
-  location_id = data.megaport_location.ndc_b1.id
+  location_id = data.megaport_location.bne_nxt1.id
   term        = 1
 }
-