+++ schema_megaport/megaport_azure_connection.go
@@ -36,8 +36,10 @@ func ResourceAzureConnectionVXCSchema() map[string]*schema.Schema {
 			Type:     schema.TypeString,
 			Computed: true,
 		},
-		"a_end": ResourcePartnerConnectionEndConfiguration(),
-		"b_end": DataVxcEndConfiguration(),
+		"a_end":                   ResourceVxcEndConfiguration(),
+		"b_end":                   DataVxcEndConfiguration(),
+		"csp_settings":            ResourceAzureConnectionCspSettings(),
+		"a_end_mcr_configuration": ResourceMcrConfigurationSettings(),
 		"created_by": {
 			Type:     schema.TypeString,
 			Computed: true,
@@ -62,7 +64,6 @@ func ResourceAzureConnectionVXCSchema() map[string]*schema.Schema {
 			Type:     schema.TypeBool,
 			Computed: true,
 		},
-		"csp_settings": ResourceAzureConnectionCspSettings(),
 		"vxc_internal_type": {
 			Type:     schema.TypeString,
 			Optional: true,
@@ -77,11 +78,6 @@ func ResourceAzureConnectionCspSettings() *schema.Schema {
 		Required: true,
 		Elem: &schema.Resource{
 			Schema: map[string]*schema.Schema{
-				"attached_to": {
-					Type:     schema.TypeString,
-					Required: true,
-					ForceNew: true,
-				},
 				"service_key": {
 					Type:     schema.TypeString,
 					Required: true,